csv.Sniffer only on first line of file.
There was an issue reported that on some file the
csv sniffer failes when looking at the first
2000 bytes of the file. It was recommended
to change to just look at the first line.
That ought to always work actually because
no meaningful header line can have only
a single cell, so there must be a separator.

See https://www.boardgamegeek.com/article/26434535#26434535