Bleed and bug-fixes.
Fixed problems with adding many copies of the
same double-sided counter spanning across
multiple boxes in the layout.

Added proper support for bleed, making it possible
to have templates with bleed areas outside of
the id rectangle.

A bug with many duplicate backside counters
sometimes being generated was fixed as a
side-effect of having to rewrite a lot
of the code for handling counter backsides.