Fixes issue #34, now remembers all text style attributes and only changes font-weight or font-style to bold or italics.
Manually merging pull request #37 by fractalpixel again, essentially, because the
2.2dev branch had significantly diverged from master and the
merge got ugly.

Hopefully everything works OK now so that both the new text style
fixes on the 2.2dev branch as well as the bug fix for #34 are
still OK. The tests look OK, but there might be some detail
lost in the noise of things that were supposed to change.