Fixes #42 by applying use translate properly.
As the translation is calculated on the original
position of what is cloned (where the use-element
would be without transform) it has to be appended
to any original transform, unlike the other transforms
that are done on some existing thing and must
be prepended. A bit of a hack now as it is late.
There are probably still ways to mess up clones by
applying combinations of transforms to the clone
or/and clone element.