Issue #46 wrong offset of templates in groups.
A template group that is in a group (instead of
being the immediate child of its layer) was
not translated correctly if it had a transform.
To reproduce this open examples/counter.svg
and move the counter template before running
the extension effect. This fix uses the top-level
group as the template, which is the expected
behaviour, not just the first parent group found.
This also means that it is easier than before
to create templates without worrying about how
many nested groups there are around the
id rectangle.
Test-cases indicates tiny changes in the
locations of images because of this. It is not
obvious if the old or new positions are better,
or why this even happened, so that might or
might not be something to fix.