Counters with rectangles as part of unit icon.
This is in preparation for hopefully soon adding
icons that are not NATO symbols with a rectangle,
but that can still use the same background unit
template as the units with NATO icons. It also
as a side-effect makes it easier to edit the
icons, which will be useful right away to add
some more standard NATO symbols.