Fixed some broken unit-tests.
Well, one test was removed, because maintaining
mock-tests is not all that productive. Like mock-tests
do it mostly tests that the mock was correctly set up,
not really that the tested function did anything right.
There are other tests in this project for testing
that the extension does what it is supposed to do.