Better error message for bad clone target
A broken clone element can not be affected
by the extension, since that will break
the algorithm for transforming the new clone target.

Possibly there is some fix for that, but now at
least the error message when it happens is
hopefully slightly less confusing
(better than just printing a stacktrace?).