Only apply ENDROW and ENDBOX after repeat
There was an issue (GitHub #113) that ENDROW and
ENDBOX are applied after every instance of a repeated
counter. The break should only be added after the
last repeated counter. This fix should do that.
At least it seems to work.

Did not test how this interacts with the inexact
repeats (e.g. RepeatMinFillRow). It is not
obvious even what is supposed to happen, or
why anyone would use an explicit ENDROW or ENDBOX
in combination with one of those repeats.